Rydda kernel
