+++ src/the_byte_of_33/include/piano_mode.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <libc/stdint.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void piano_mode(void);
+
+#ifdef __cplusplus
+}
+#endif
